backend/requirements.txt
fastapi>=0.111.0
pyjwt>=2.8.0
passlib>=1.7.4
beanie>=1.26.0
pydantic-settings>=2.2.1
python-jose>=3.3.0
pytest>=8.2.1
mongomock>=4.1.2
mongomock_motor>=0.0.29
pytest-asyncio>=0.23.7
